Refactor `file_handler` ->`conf/handler/file.py`
